pyproject.toml
[project]
name = "x-daily-digest"
version = "1.0.0"
description = "Personal X/Twitter email digest of tweets from accounts you follow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tweepy>=4.14.0",
    "resend>=0.7.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pytz>=2024.1",
]

[project.scripts]
x-digest = "src.main:main"

